Utils

Physical Constants

WaveSpec.PhysicalConstants.g — Constant

g

Standard gravity (m/s²)

WaveSpec.PhysicalConstants.ν_water — Constant

ν_water

Kinematic viscosity of water (m²/s) at 15°C.

WaveSpec.PhysicalConstants.ρ_air — Constant

ρ_air

Density of air (kg/m³) at sea level.

WaveSpec.PhysicalConstants.ρ_water — Constant

ρ_water

Density of seawater (kg/m³). Standard value at 15°C and 3.5% salinity.

WaveSpec.PhysicalConstants.σ_water — Constant

σ_water

Surface tension of water-air interface (N/m).

Integration

WaveSpec.Integration.IntegrateGaussQuad — Method

IntegrateGaussQuad(f, order, a, b, n)

Integrates function f from a to b using a composite Gauss-Legendre rule of order 'order' and and n sub-intervals (panels). The mapping used is: x = (b-a)/2 * ξ + (a+b)/2 dx = (b-a)/2 * dξ

WaveSpec.Integration.IntegrateTrapezoidal — Method

IntegrateTrapezoidal(f; a, b, n)

Integrates function f from a to b using the Composite Trapezoidal Rule with n sub-intervals (panels).

Truncation

Signal