Spectral Spreading

WaveSpec.SpectralSpreading.DiscreteSpectralSpreading — Type

DiscreteSpectralSpreading

The discrete realization of a frequency spectrum. Represents the energy distribution across a finite set of frequency bins.

WaveSpec.SpectralSpreading.DiscreteSpectralSpreading — Method

DiscreteSpectralSpreading(shape::RegularWave; mess=true)

Specialized constructor for RegularWave spectrum type.

WaveSpec.SpectralSpreading.DiscreteSpectralSpreading — Method

DiscreteSpectralSpreading(shape, strategy, fmin, fmax, nf)

Continuous abstract spectrum discretised according to the selected abstract sampling strategy, between fmin and fmax with nf bins.

WaveSpec.SpectralSpreading.get_Hs — Method

get_Hs(spec::DiscreteSpectralSpreading)

Returns the corrected significant wave height Hs = 4√m₀ computed from the discrete bins.

WaveSpec.SpectralSpreading.get_amplitude — Method

get_amplitude(spec::DiscreteSpectralSpreading, idx::Int)

Returns corrected wave amplitude A(fᵢ) for the specified bin index.

WaveSpec.SpectralSpreading.get_amplitudes — Method

get_amplitudes(spec::DiscreteSpectralSpreading, r::AbstractUnitRange{Int})

Returns corrected wave amplitudes A(fᵢ) for the specified bin range.

WaveSpec.SpectralSpreading.get_bandwidths — Method

get_bandwidths(spec::DiscreteSpectralSpreading)

Returns the bandwidths of each discrete bin.

WaveSpec.SpectralSpreading.get_central_frequencies — Method

get_central_frequencies(spec::DiscreteSpectralSpreading)

Returns the central frequencies of each discrete bin.

WaveSpec.SpectralSpreading.get_central_frequency — Method

get_central_frequency(spec::DiscreteSpectralSpreading, idx::Int)

Returns the central frequency of the bin at the given index.

WaveSpec.SpectralSpreading.get_densities — Method

get_densities(spec::DiscreteSpectralSpreading, idx0::Int=1, idx1::Int=length(spec.fᵢ))

Returns corrected spectral densities S(fᵢ) for the specified bin range.

WaveSpec.SpectralSpreading.get_densities — Method

get_densities(spec::DiscreteSpectralSpreading)

Returns corrected spectral densities S(fᵢ) for all bins.

WaveSpec.SpectralSpreading.get_density — Method

get_density(spec::DiscreteSpectralSpreading, idx::Int)

Returns corrected spectral density S(fᵢ) for the specified bin index.

WaveSpec.SpectralSpreading.get_energies — Method

get_energies(spec::DiscreteSpectralSpreading, r::AbstractUnitRange{Int})

Returns corrected wave energies E(fᵢ) for the specified bins range.

WaveSpec.SpectralSpreading.get_energy — Method

get_energy(spec::DiscreteSpectralSpreading, idx::Int)

Returns corrected wave energy E(fᵢ) for the specified bin index.

WaveSpec.SpectralSpreading.get_frequencies — Method

get_frequencies(spec::DiscreteSpectralSpreading)

Returns the discrete frequency bins and their widths.

WaveSpec.SpectralSpreading.get_frequency — Method

get_frequency(spec::DiscreteSpectralSpreading, idx::Int)

Returns the edge frequency at the given index.

WaveSpec.SpectralSpreading.get_frequency — Method

get_frequency(spec::DiscreteSpectralSpreading, f_target::Real)

Returns the edge frequency closest to f_target.

WaveSpec.SpectralSpreading.get_integrated_energies — Method

get_integrated_energies(spec::DiscreteSpectralSpreading; nfine = 200)

Returns the energies per bin by integrating the continuous spectrum inside each bin (-> continuous spectrum-wise energy inside discrete-wise bins).

WaveSpec.SpectralSpreading.get_moment — Method

get_moment(spec::DiscreteSpectralSpreading, n::Int)

Calculates the n-th corrected discrete spectral moment: mₙ = Σ (fⁿ * S * df)

WaveSpec.SpectralSpreading.get_spectral_index — Method

get_spectral_index(spec::DiscreteSpectralSpreading, f_range::AbstractRange{<:Real})

Returns a UnitRange of bin indexes that cover the provided frequency range. Useful for extracting a "slice" of the spectrum (e.g., around the peak).

WaveSpec.SpectralSpreading.get_spectral_index — Method

get_spectral_index(spec::DiscreteSpectralSpreading, f_target::Real)

Returns the index (1 to nf-1) of the bin containing the frequency f_target.